Create a Formatters Directory and Move Formatting Functions
Related to #264
